Support running as local instead of dynamic user
The `earlyoom` service unit on the Redhat family contains:

```ini
[Service]
DynamicUser=true
```
If local_user is specified as true (not the default) a local
user `earlyoom` will be created and systemd will be instructed to use
that user.

On Debian based earlyoom runs as root as this parameter will be ignored.